Rename `testing` packages to `protocoltest`
