+++ lib/srv/db/access_test.go
@@ -77,7 +77,7 @@ import (
 	"github.com/gravitational/teleport/lib/srv"
 	"github.com/gravitational/teleport/lib/srv/alpnproxy"
 	alpncommon "github.com/gravitational/teleport/lib/srv/alpnproxy/common"
-	cassandra "github.com/gravitational/teleport/lib/srv/db/cassandra/testing"
+	cassandra "github.com/gravitational/teleport/lib/srv/db/cassandra/protocoltest"
 	"github.com/gravitational/teleport/lib/srv/db/clickhouse"
 	"github.com/gravitational/teleport/lib/srv/db/cloud"
 	"github.com/gravitational/teleport/lib/srv/db/common"
@@ -91,9 +91,9 @@ import (
 	"github.com/gravitational/teleport/lib/srv/db/opensearch"
 	"github.com/gravitational/teleport/lib/srv/db/postgres"
 	redisprotocol "github.com/gravitational/teleport/lib/srv/db/redis/protocol"
-	redis "github.com/gravitational/teleport/lib/srv/db/redis/testing"
+	redis "github.com/gravitational/teleport/lib/srv/db/redis/protocoltest"
 	"github.com/gravitational/teleport/lib/srv/db/snowflake"
-	spanner "github.com/gravitational/teleport/lib/srv/db/spanner/testing"
+	spanner "github.com/gravitational/teleport/lib/srv/db/spanner/protocoltest"
 	"github.com/gravitational/teleport/lib/srv/db/sqlserver"
 	"github.com/gravitational/teleport/lib/srv/discovery/fetchers/db"
 	"github.com/gravitational/teleport/lib/tlsca"
