+++ lib/srv/db/audit_test.go
@@ -37,7 +37,7 @@ import (
 	"github.com/gravitational/teleport/lib/events/eventstest"
 	"github.com/gravitational/teleport/lib/srv/db/common"
 	"github.com/gravitational/teleport/lib/srv/db/postgres"
-	redis "github.com/gravitational/teleport/lib/srv/db/redis/testing"
+	redis "github.com/gravitational/teleport/lib/srv/db/redis/protocoltest"
 )
 
 // TestAuditPostgres verifies proper audit events are emitted for Postgres
