+++ lib/srv/db/redis/protocoltest/test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package testing
+package protocoltest
 
 import (
 	"context"
