+++ lib/srv/db/spanner_test.go
@@ -35,7 +35,7 @@ import (
 	"github.com/gravitational/teleport/lib/defaults"
 	libevents "github.com/gravitational/teleport/lib/events"
 	"github.com/gravitational/teleport/lib/srv/db/common"
-	spanner "github.com/gravitational/teleport/lib/srv/db/spanner/testing"
+	spanner "github.com/gravitational/teleport/lib/srv/db/spanner/protocoltest"
 )
 
 func TestAccessSpanner(t *testing.T) {
